# src/za.co.grindrodbank.a3s/wwwroot/openapi-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "description" : "API Definition for the A3S. This service allows authentication, authorisation and accounting.",
    "title" : "A3S",
    "version" : "1.0.5"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
    "/users" : {
      "get" : {
        "description" : "Search for users.",
        "operationId" : "listUsers",
        "parameters" : [ {
          "description" : "Whether to fill in the teams member field",
          "explode" : true,
          "in" : "query",
          "name" : "teams",
          "required" : false,
          "schema" : {
            "default" : false,
            "type" : "boolean"
          },
          "style" : "form"
        }, {
          "description" : "Whether to fill in the roles member field",
          "explode" : true,
          "in" : "query",
          "name" : "roles",
          "required" : false,
          "schema" : {
            "default" : false,
            "type" : "boolean"
          },
          "style" : "form"
        }, {
          "description" : "Whether to fill in the functions member field",
          "explode" : true,
          "in" : "query",
          "name" : "functions",
          "required" : false,
          "schema" : {
            "default" : false,
            "type" : "boolean"
          },
          "style" : "form"
        }, {
          "description" : "If this field is set, translate all applicable fields to a specific locale\n",
          "explode" : true,
          "in" : "query",
          "name" : "locale",
          "required" : false,
          "schema" : {
            "default" : "en",
            "example" : "[\"en-za\",\"en-us\",\"af-za\",\"en\"]",
            "maxLength" : 5,
            "minLength" : 2,
            "type" : "string"
          },
          "style" : "form"
        }, {
          "description" : "The page to view.",
          "explode" : true,
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "minimum" : 0,
            "type" : "integer"
          },
          "style" : "form"
        }, {
          "description" : "The size of a page.",
          "explode" : true,
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 10,
            "format" : "int32",
            "maximum" : 20,
            "minimum" : 1,
            "type" : "integer"
          },
          "style" : "form"
        }, {
          "description" : "A search query filter on the User's name.",
          "example" : "FilterByEntityType",
          "explode" : true,
          "in" : "query",
          "name" : "filterName",
          "required" : false,
          "schema" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "style" : "form"
        }, {
          "description" : "A search query filter on the User's username",
          "explode" : true,
          "in" : "query",
          "name" : "filterUsername",
          "required" : false,
          "schema" : {
            "type" : "string"
          },
          "style" : "form"
        }, {
          "description" : "a comma separated list of fields in their sort order. Ascending order is assumed. Append desc after a field to indicate descending order.",
          "explode" : false,
          "in" : "query",
          "name" : "orderBy",
          "required" : false,
          "schema" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "style" : "form"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/User"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK",
            "headers" : {
              "X-Pagination" : {
                "description" : "Pagination header.",
                "explode" : false,
                "schema" : {
                  "$ref" : "#/components/schemas/PaginationHeaderResponse"
                },
                "style" : "simple"
              }
            }
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Bad Request."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Not authenticated."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Forbidden -  You are not authorized to access the list of users."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "User list not found."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "An unexpected error occurred."
          }
        },
        "summary" : "Search for users.",
        "tags" : [ "User" ]
      },
      "post" : {
        "description" : "Create a new User.",
        "operationId" : "createUser",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserSubmit"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserSubmit"
                }
              }
            },
            "description" : "Successful. User created."
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Invalid parameters."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Not authenticated."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Forbidden - Not authorized to create a user."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Related user entity (such as role or team) not found."
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Non-Processible entity. The request was correctly structured, but some business rules were violated, preventing the user creation."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "An unexpected error occurred."
          }
        },
        "summary" : "Create a User.",
        "tags" : [ "User" ]
      }
    },
...
    "/users/{userId}/changePassword" : {
      "put" : {
        "description" : "Change a user password.",
        "operationId" : "changeUserPassword",
        "parameters" : [ {
          "description" : "The UUID of the user.",
          "explode" : false,
          "in" : "path",
          "name" : "userId",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UserId"
          },
          "style" : "simple"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserPasswordChangeSubmit"
              }
            }
          }
        },
        "responses" : {
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Bad Request."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Not authenticated."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Forbidden - You are not authorized update user passwords."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "User not found."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "An unexpected error occurred."
          }
        },
        "summary" : "Change a user password.",
        "tags" : [ "User" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "UserId" : {
        "format" : "uuid",
        "type" : "string"
      },
      "User" : {
        "description" : "A user\n",
        "example" : {
          "phoneNumber" : "phoneNumber",
          "isDeleted" : true,
          "teams" : [ {
            "subRealmId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
            "userIds" : [ "046b6c7f-0b8a-43b9-b35d-6489e6daee91", "046b6c7f-0b8a-43b9-b35d-6489e6daee91" ],
            "name" : "name",
            "description" : "description",
            "termsOfServiceId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
            "uuid" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
            "dataPolicyIds" : [ null, null ],
            "teamIds" : [ "046b6c7f-0b8a-43b9-b35d-6489e6daee91", "046b6c7f-0b8a-43b9-b35d-6489e6daee91" ]
          }, {
            "subRealmId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
            "userIds" : [ "046b6c7f-0b8a-43b9-b35d-6489e6daee91", "046b6c7f-0b8a-43b9-b35d-6489e6daee91" ],
            "name" : "name",
            "description" : "description",
            "termsOfServiceId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
            "uuid" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
            "dataPolicyIds" : [ null, null ],
            "teamIds" : [ "046b6c7f-0b8a-43b9-b35d-6489e6daee91", "046b6c7f-0b8a-43b9-b35d-6489e6daee91" ]
          } ],
          "surname" : "surname",
          "roles" : [ {
            "subRealmId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
            "roleIds" : [ null, null ],
            "userIds" : [ "046b6c7f-0b8a-43b9-b35d-6489e6daee91", "046b6c7f-0b8a-43b9-b35d-6489e6daee91" ],
            "name" : "name",
            "description" : "description",
            "functionIds" : [ "046b6c7f-0b8a-43b9-b35d-6489e6daee91", "046b6c7f-0b8a-43b9-b35d-6489e6daee91" ],
            "uuid" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
          }, {
            "subRealmId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
            "roleIds" : [ null, null ],
            "userIds" : [ "046b6c7f-0b8a-43b9-b35d-6489e6daee91", "046b6c7f-0b8a-43b9-b35d-6489e6daee91" ],
            "name" : "name",
            "description" : "description",
            "functionIds" : [ "046b6c7f-0b8a-43b9-b35d-6489e6daee91", "046b6c7f-0b8a-43b9-b35d-6489e6daee91" ],
            "uuid" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
          } ],
          "name" : "name",
          "avatar" : "avatar",
          "ldapAuthenticationModeId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
          "uuid" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
          "email" : "email",
          "username" : "username"
        },
        "properties" : {
          "uuid" : {
            "format" : "uuid",
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "surname" : {
            "type" : "string"
          },
          "username" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "phoneNumber" : {
            "nullable" : true,
            "type" : "string"
          },
          "avatar" : {
            "nullable" : true,
            "type" : "string"
          },
          "isDeleted" : {
            "description" : "True if this user is considered deleted.",
            "type" : "boolean"
          },
          "ldapAuthenticationModeId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "roles" : {
            "items" : {
              "$ref" : "#/components/schemas/Role"
            },
            "type" : "array"
          },
          "teams" : {
            "items" : {
              "$ref" : "#/components/schemas/Team"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "UserPasswordChangeSubmit" : {
        "description" : "Used to change a user's password.\n",
        "example" : {
          "oldPassword" : "oldPassword",
          "newPassword" : "newPassword",
          "uuid" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
          "newPasswordConfirmed" : "newPasswordConfirmed"
        },
        "properties" : {
          "uuid" : {
            "format" : "uuid",
            "type" : "string"
          },
          "newPassword" : {
            "type" : "string"
          },
          "newPasswordConfirmed" : {
            "type" : "string"
          },
          "oldPassword" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ErrorResponse" : {
        "description" : "Error response object to give more info with HTTP status code",
        "properties" : {
          "message" : {
            "type" : "string"
          }
        },
        "required" : [ "message" ],
        "type" : "object"
      },
      "PaginationHeaderResponse" : {
        "description" : "This is the standard response for requests that have paginated collections",
        "properties" : {
          "total" : {
            "description" : "The total number of pages in the result set",
            "format" : "int32",
            "minimum" : 1,
            "type" : "integer"
          },
          "count" : {
            "description" : "The total number of results in the result set",
            "format" : "int32",
            "minimum" : 1,
            "type" : "integer"
          },
          "size" : {
            "default" : 10,
            "description" : "The number items to include in a page of results. The page size for the result set",
            "format" : "int32",
            "maximum" : 20,
            "minimum" : 1,
            "type" : "integer"
          },
          "current" : {
            "description" : "The position of the page in the paged result set that is being returned",
            "format" : "int32",
            "minimum" : 1,
            "type" : "integer"
          },
          "first" : {
            "description" : "The link to the first page of results containing [size] results. This link includes the page, size, filter, orderBy and fields query parameters.",
            "format" : "url",
            "type" : "string"
          },
          "last" : {
            "description" : "The link to the last page of results containing [size] results. This link includes the page, size, filter, orderBy and fields query parameters. If this is the first page in the result set then this will be [null].",
            "format" : "url",
            "type" : "string"
          },
          "prev" : {
            "description" : "The link to the previous page of results containing [size] results. This link includes the page, size, filter, orderBy and fields query parameters. If this is the first page in the result set then this will be [null].",
            "format" : "url",
            "type" : "string"
          },
          "next" : {
            "description" : "The link to the next page of results containing [size] results. This link includes the page, size, filter, orderBy and fields query parameters. If this is the last page in the result set then this will be [null].",
            "format" : "url",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    }
  }
}